refactor(backend): setup scalable codebase
