+++ database/db.js
@@ -1,5 +1,7 @@
-require("dotenv").config();
-import { Pool } from "pg";
+import pkg from "pg";
+const { Pool } = pkg;
+import dotenv from "dotenv";
+dotenv.config();
 
 const pool = new Pool({
   connectionString: process.env.DATABASE_URL,
