+++ package.json
@@ -1,6 +1,7 @@
 {
   "name": "js-project",
   "version": "1.0.0",
+  "type": "module",
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
